chore(editor-state): move compiler related files into the compiler features
