+++ packages/editor/packages/editor-state/src/features/config-compiler/extractConfigBody.ts
