+++ packages/editor/packages/editor-state/src/features/config-compiler/isPlainObject.ts
