+++ packages/editor/packages/editor-state/src/features/program-compiler/compiler.ts
@@ -1,9 +1,9 @@
 import { StateManager } from '@8f4e/state-manager';
 
-import { EventDispatcher } from '../types';
-import { error, log } from '../impureHelpers/logger/logger';
+import { EventDispatcher } from '../../types';
+import { error, log } from '../../impureHelpers/logger/logger';
 
-import type { CodeBlockGraphicData, State } from '../types';
+import type { CodeBlockGraphicData, State } from '../../types';
 
 /**
  * Converts code blocks into separate arrays for modules and functions, sorted by creationIndex.
