+++ packages/editor/packages/editor-state/src/index.ts
@@ -9,8 +9,8 @@ import codeBlockCreator from './effects/codeBlocks/codeBlockCreator';
 import codeBlockDragger from './effects/codeBlocks/codeBlockDragger';
 import codeBlockNavigation from './effects/codeBlockNavigation';
 import demoModeNavigation from './effects/demoModeNavigation';
-import compiler from './effects/compiler';
-import configEffect from './effects/config';
+import compiler from './features/program-compiler/compiler';
+import configEffect from './features/config-compiler/config';
 import contextMenu from './effects/menu/contextMenu';
 import graphicHelper from './effects/codeBlocks/graphicHelper';
 import editorSettings from './effects/editorSettings';
